refactor(fs-bridge): move BridgeUnsupportedOperation to a new errors module
* Created a new `errors.ts` file to house the `BridgeUnsupportedOperation` class.
* Updated import paths in `assertions.ts` and `define.ts` to reflect the new location.
* Ensured that the `index.ts` file exports the new error class appropriately.
