+++ packages/fs-bridge/src/index.ts
@@ -1,5 +1,11 @@
-export { assertCapability, BridgeUnsupportedOperation } from "./assertions";
+export {
+  assertCapability,
+} from "./assertions";
+
 export { defineFileSystemBridge } from "./define";
+
+export * from "./errors";
+
 export type {
   FileSystemBridge,
   FileSystemBridgeOperations,
